relocate the sync dir
