+++ src/plugins/dashboard/public/application/embeddable/direct_query_sync/_dashboard_direct_query_sync.scss
