+++ src/plugins/dashboard/public/application/embeddable/direct_query_sync/dashboard_direct_query_sync.test.tsx
