+++ src/plugins/dashboard/public/application/embeddable/direct_query_sync/dashboard_direct_query_sync.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 import { EuiCallOut, EuiLink, EuiLoadingSpinner, EuiText } from '@elastic/eui';
 import { i18n } from '@osd/i18n';
 import { DirectQueryLoadingStatus } from '../../../../../data_source_management/public';
-import { EMR_STATES, intervalAsMinutes } from '../../utils/direct_query_sync/direct_query_sync';
+import { EMR_STATES, intervalAsMinutes } from './direct_query_sync';
 import './_dashboard_direct_query_sync.scss';
 
 export interface DashboardDirectQuerySyncProps {
