+++ src/plugins/dashboard/public/application/embeddable/direct_query_sync/direct_query_sync.ts
